# Feat/cache queries