--- a/pyproject.toml
+++ b/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "plexosdb"
-version = "0.0.6"
+version = "0.0.7"
 readme = "README.md"
 license = {file = "LICENSE.txt"}
 keywords = []