--- a/tests/data/plexosdb.xml
+++ b/tests/data/plexosdb.xml
@@ -82,7 +82,7 @@
 	<t_object>
 		<object_id>3</object_id>
 		<class_id>2</class_id>
-		<name>SolarPV01</name>
+		<name>SolarPV02</name>
 		<category_id>97</category_id>
 		<GUID>40d15a07-8ccc-460e-919a-ec8e211899a8</GUID>
 	</t_object>