--- a/tests/test_plexosdb_sqlite.py
+++ b/tests/test_plexosdb_sqlite.py
@@ -6,7 +6,6 @@ from plexosdb.sqlite import PlexosSQLite
 from sqlite3 import IntegrityError
 from collections.abc import Generator
 from pathlib import Path
-from tempfile import TemporaryDirectory
 
 DB_FILENAME = "plexosdb.xml"
 
@@ -17,10 +16,9 @@ def db_empty() -> "PlexosSQLite":
 
 
 @pytest.fixture
-def db(data_folder: Path) -> Generator[PlexosSQLite, None, None]:
-    tmp_dir = TemporaryDirectory()  # dir=data_folder.parent)
+def db(data_folder: Path, tmp_path: Path) -> Generator[PlexosSQLite, None, None]:
     xml_fname = data_folder / DB_FILENAME
-    xml_copy = Path(tmp_dir.name) / f"copy_{DB_FILENAME}"
+    xml_copy = tmp_path / f"copy_{DB_FILENAME}"
     shutil.copy(xml_fname, xml_copy)
     db = PlexosSQLite(xml_fname=str(xml_copy))
     yield db