Cache results from _get_id
In the run of:
`r2x -i temp_data/oct1024_Con_NG  --input-model=reeds-US --output-model=plexos --weather-year=2012 -o .out/$CASETS --year 2035 --plugins pcm_defaults emission_cap break_gens imports --debug --upgrade --flags daily-budgets=true`

In this case:
_get_id is called 434408 times
there are only 5397 unique calls of _get_id

Caching _get_id results leads to the performance gains:
plexos_exporter: 47.3s -> 19.7s
_get_id cumulative: 28.8s -> 1.7s