@JensZack 2024-11-21
I would like to clean up the cacheing mechanism to only use class/object name, for id queries in the object table. This also raises questions about what uniqueness can be enforced in other tables where _get_id is used repeatedly.